correct @RequestParam in controller

+++ src/main/java/org/da1/paymentservice/dtos/InitiatePaymentDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class InitiatePaymentDTO {
-       Double amount;
+       Long amount;
        String orderId;
        String phoneNumber;
        String name;
