+++ src/main/java/org/da1/paymentservice/paymentgateways/IPaymentGateway.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Component;
 @Component
 public interface IPaymentGateway {
 
-    String createStandardPaymentLink(Double amount,String orderId,
+    String createStandardPaymentLink(Long amount,String orderId,
                                      String phoneNumber,
                                      String name,
                                      String email);
