+++ src/main/java/org/da1/paymentservice/services/IPaymentService.java
@@ -1,7 +1,7 @@
 package org.da1.paymentservice.services;
 
 public interface IPaymentService {
-    String getPaymentLink(Double amount,String orderId,
+    String getPaymentLink(Long amount,String orderId,
                           String phoneNumber,
                           String name,
                           String email);
